requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==4.2
django-cors-headers==4.0.0
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.14.0
Flask==3.0.3
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.4.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
openapi-codec==1.3.2
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
pillow==9.3.0
pipenv==2023.2.18
platformdirs==3.11.0
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
reportlab==4.0.4
requests==2.32.3
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.20.0
virtualenv-clone==0.5.7
Werkzeug==3.0.3
# CinetPay==0.0.1